+++ CHANGELOG.md
@@ -53,6 +53,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added endpoint `set_excess_energy_heating_limit_night()` to the heating circuit
 - Added endpoint `get_excess_energy_cooling_limit_night()` to the heating circuit
 - Added endpoint `set_excess_energy_cooling_limit_night()` to the heating circuit
+- Added endpoint `get_pump_state()` to the heating circuit
+- Added endpoint `get_mixer_position()` to the heating circuit
 - Added endpoint `get_excess_energy_mode()` to the heating circuit
 - Added endpoint `get_use_excess_energy()` to the heating circuit
 - Added endpoint `set_use_excess_energy()` to the heating circuit
@@ -65,6 +67,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added endpoint `get_min_runtime_excess_energy()` to the external heat source
 - Added endpoint `set_min_runtime_excess_energy()` to the external heat source
 - Added endpoint `get_excess_energy_active()` to the photovoltaics
+- Added endpoint `get_mixer_position()` to passive cooling
 
 ## [2.11.0] - 2026-04-13